perf ColorPalette 드래그 좌우 이동 이벤트 재구현
