+++ src/components/molecules/colorPalette/ColorPalette.tsx
@@ -2,47 +2,18 @@ import ColorCircle from '@/components/atoms/colorCircle/ColorCircle';
 import React, { useEffect, useRef, useState } from 'react';
 import useColorPalette from './ColorPalette.hook';
 import { ColorPaletteProps } from './colorPalette.types';
+import { throttle } from '@/utils/common';
 
 const ColorPalette = ({ handleSetColor }: ColorPaletteProps) => {
-  const { bgColors, borderColors, handleClickColorCircle } =
-    useColorPalette(handleSetColor);
-  const colorPaletteRef = useRef<HTMLDivElement>(null);
-
-  const [isDrag, setIsDrag] = useState(false);
-  const [startX, setStartX] = useState(0);
-
-  const onDragStart = (e: React.MouseEvent) => {
-    if (!colorPaletteRef.current) return;
-    e.preventDefault();
-    setIsDrag(true);
-    setStartX(e.pageX + colorPaletteRef.current?.scrollLeft);
-  };
-  const onDragEnd = () => {
-    setIsDrag(false);
-  };
-
-  const onDragMove = (e: React.MouseEvent) => {
-    if (isDrag && colorPaletteRef.current) {
-      colorPaletteRef.current.scrollLeft = startX - e.pageX;
-    }
-  };
-
-  const throttle = <T extends (...args: any[]) => any>(fn: T, ms: number) => {
-    let isExcute = false;
-
-    return (...args: Parameters<T>) => {
-      if (!isExcute) {
-        isExcute = true;
-        setTimeout(() => {
-          // console.log(...args);
-          fn(...args);
-          isExcute = false;
-        }, ms);
-      }
-    };
-  };
-
-  const onThrottleDragMove = throttle(onDragMove, 25);
+  const {
+    bgColors,
+    borderColors,
+    colorPaletteRef,
+    handleClickColorCircle,
+    onDragStart,
+    onDragEnd,
+    onThrottleDragMove,
+  } = useColorPalette(handleSetColor);
 
   return (
     <div
